Fix compilation errors - 22 build issues resolved
Fixed all compilation errors from Android build:
- Import msg_super macro in media_item.rs and ui_table_view.rs
- Change HostConstant::Double to Custom with f64 allocation
- Fix Cow<str> pattern issues by using .as_ref()
- Fix temporary value lifetime issue in replace_case_insensitive
- Import CallFromHost trait in cxxabi.rs
- Fix destructor call signature (tuple instead of array)
- Fix timestamp types (i32 instead of i64 for time_t and tv_nsec)
- Fix st_blocks type cast (u64)
- Simplify borrow_mut panic message (removed class field access)
- Fix unused variable warning (_matched)

+++ src/libc/posix_io/stat.rs
@@ -116,7 +116,7 @@ fn fstat_inner(env: &mut Environment, fd: FileDescriptor, buf: MutPtr<stat>) -> 
             
             // Block info
             stat_data.st_blksize = 4096;
-            stat_data.st_blocks = (stat_data.st_size + 511) / 512;
+            stat_data.st_blocks = ((stat_data.st_size + 511) / 512) as u64;
         }
         GuestFile::Directory => {
             stat_data.st_mode |= S_IFDIR;
@@ -139,8 +139,8 @@ fn fstat_inner(env: &mut Environment, fd: FileDescriptor, buf: MutPtr<stat>) -> 
     let now = std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
         .unwrap();
-    let now_secs = now.as_secs() as i64;
-    let now_nsecs = now.subsec_nanos() as i64;
+    let now_secs = now.as_secs() as i32;
+    let now_nsecs = now.subsec_nanos() as i32;
     
     stat_data.st_atimespec = timespec {
         tv_sec: now_secs,
